Emit 'done' when download really completes
If fast-download is used to write directly to a file, it emits the 'end'
event when the download completes and not when file has completely
written. This could be an issue if the file is read in the 'end' event
handler because it can be incomplete.
The 'done' event is emitted always when the download is really
completed.